backend/pyproject.toml
[project]
name = "edtronaut-ai-coworker"
version = "0.1.0"
description = "AI Coworker for business simulation training"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # ── Core Framework ──
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    # ── gRPC ──
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    # ── LLM & AI ──
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.40",
    "langfuse>=2.0.0",
    "openai>=1.10.0",
    # ── Databases ──
    "redis>=5.0.0",
    "qdrant-client[grpc]>=1.7.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # ── Auth & Security ──
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # ── Utilities ──
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coworker_api"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
